`createVertexBuffer` の型変更と型キャストの追加
`Mesh.h` と `Mesh.cpp` において、`createVertexBuffer` メソッドの `stride` 引数の型を `uint32_t` から `uint8_t` に変更しました。この変更により、メモリ使用量の最適化が図られています。

さらに、`Mesh::addSurfaceFromArrays` メソッドでは、`vertexCount`、`vertexElementSize`、および `indexCount` に対して適切な型キャストを追加し、型の整合性を確保しました。

また、`Mesh::createVertexBuffer` メソッド内で `size` 引数に `static_cast<uint32_t>` を適用し、`VertexBuffer::create` 呼び出し時の型の明示的な統一を行いました。

+++ engine/include/ocf/3d/Mesh.h
@@ -60,7 +60,7 @@ protected:
 
     VertexBuffer* createVertexBuffer(uint64_t format, uint32_t vertexCount,
                                      const std::array<uint32_t, ArrayType::ArrayMax>& offsets,
-                                     uint32_t stride, const void* data, size_t size);
+                                     uint8_t stride, const void* data, size_t size);
 
     IndexBuffer* createIndexBuffer(uint32_t indexCount, const void* data, size_t size);
